requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pydantic
PyJWT
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
gunicorn
supabase
python-dotenv
google-generativeai